+++ {{cookiecutter.repo_name}}/project/accounts/__init__.py
@@ -1,0 +1,5 @@
+"""
+Common account app for login/registration
+
+Enabled and configured from mod 'accounts'
+"""