+++ {{cookiecutter.repo_name}}/project/accounts/admin.py
@@ -1,11 +1,11 @@
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 
-from .models import UserInfo
+from .models import UserProfile
 
-class UserInfoAdmin(admin.ModelAdmin):
+class UserProfileAdmin(admin.ModelAdmin):
     list_display = ('user', 'company', 'function',)
     raw_id_fields = ['user']
     search_fields = ('user__username', 'company',)
 
-admin.site.register(UserInfo, UserInfoAdmin)
+admin.site.register(UserProfile, UserProfileAdmin)