+++ {{cookiecutter.repo_name}}/project/mods_available/accounts/settings.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
-INSTALLED_APPS = add_to_tuple(INSTALLED_APPS, 'project.accounts', 'registration')
+INSTALLED_APPS = add_to_tuple(INSTALLED_APPS, 
+    'project.accounts',
+    'registration',
+)
 
 ACCOUNT_ACTIVATION_DAYS = 7
 DEFAULT_FROM_EMAIL = 'noreply@dummy.com'
@@ -9,3 +12,7 @@ LOGIN_REDIRECT_URL = "/"
 
 DEFAULT_ACTIVE_SUBSCRIBE_EMAILS = []
 CONTACT_FORM_EMAILS = []
+
+# Wether we send email with mixed content type plain/text+html (True) or just 
+# plain/text (False)
+REGISTRATION_EMAIL_HTML = False