+++ {{cookiecutter.repo_name}}/project/mods_available/contact_form/__init__.py
@@ -1,5 +1,6 @@
 """
 A simple contact form that is more of a standard template than a full-blown application. You can modify it according to your requirements in its ``apps/contact_form/`` directory. Its HTML rendering is managed by `crispy_forms`_ based on a customized layout.
 
-Depends on `recaptcha`_ and `crispy_forms`_ mods.
+.. note::
+   Depends on `recaptcha`_ and `crispy_forms`_ mods.
 """